Corrected a few things. Found more errors.
The SuggStats.py file is correctly reading in data from the .csv file
and the fasta file. If a variable is not found in the .csv file, it
causes the script to fail. I need to add a check to see if the variable
is in the .csv file and if not, set it to a default value.