Lots of updates.
I have renamed the alt files to the original names.
anova.py has been heavily modified and simplified so it can be used in
the stats.py file.
stats.py file has been created, and it is a work in progress.
At this point, it seems to handle the anova.py and power_analysis.py
files.