Complete.
All of the functions of the stats.py file have been implemented.
There are somet hings that I would like to add, but for the sake of the
project, I will refrain.

I will probably continue to work on this project in my free time,
but for now, I am done with the project.